requirements.txt
flask==3.0.3
playwright==1.49.1
python-dotenv==1.0.1
browser-use
langchain-groq